Remove duplicate util classes
